requirements.txt
django
djangorestframework
django-extensions
vobject
chardet
surlex
sorl-thumbnail
django-newsletter